+++ packages/magento-cart/components/CartAdded/index.ts
@@ -1,2 +1,3 @@
 export * from './CartAdded.gql'
 export * from './Crosssells.gql'
+export * from './useCrosssellItems'
